修改檔案結構
