+++ app/javascript/src/footer/text/AboutUs.jsx
@@ -1,15 +1,17 @@
 import React from 'react'
 
-function AboutUs() {
+const AboutUs = () => {
   return (
-    [
-        '關於',
-        '關於我們',
-        '挖貝徽章認證',
-        '贊助保障規範',
-        '會員紅利介紹',
-        '提案服務說明'
-    ]
+    <>
+        [
+            '關於',
+            '關於我們',
+            '挖貝徽章認證',
+            '贊助保障規範',
+            '會員紅利介紹',
+            '提案服務說明'
+        ]
+    </>
   )
 }
