+++ app/javascript/src/homepageMain/main/banner.css
