+++ app/javascript/src/homepageMain/main/components/FundraisingItem.jsx
