+++ app/javascript/src/homepageMain/main/components/ProgressBar.jsx
