+++ app/views/home/index.html.erb
@@ -1,1 +1,1 @@
-<%= javascript_pack_tag 'Wubay' %>
+<%= javascript_pack_tag 'HomePage' %>
